+++ src/App.js
@@ -1,18 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import './styles/App.css';
-import Main from './components/Main.js';
-import OfferHelp from './components/OfferHelp.js';
-import Dashboard from './components/Dashboard.js';
-import FAQ from './components/FAQ.js';
-import Impressum from './components/Impressum.js';
-import Signup from './components/Signup';
-import AskForHelp from './components/AskForHelp';
-import Overview from './components/Overview';
-import Success from './components/Success';
+import Main from './views/Main.js';
+import OfferHelp from './views/OfferHelp.js';
+import Dashboard from './views/Dashboard.js';
+import FAQ from './views/FAQ.js';
+import Impressum from './views/Impressum.js';
+import Signup from './views/Signup';
+import AskForHelp from './views/AskForHelp';
+import Overview from './views/Overview';
+import Success from './views/Success';
 import withFirebaseAuth from 'react-with-firebase-auth';
 import fb from './firebase';
-import SuccessOffer from './components/SuccessOffer';
-import DSGVO from './components/DSGVO';
+import SuccessOffer from './views/SuccessOffer';
+import DSGVO from './views/DSGVO';
 import CookieConsent from 'react-cookie-consent';
 import MenuIcon from '@material-ui/icons/Menu';
 import {
@@ -23,9 +23,9 @@ import {
 } from 'react-router-dom';
 import Sidebar from "./components/Sidebar";
 import DesktopMenu from './components/DesktopMenu';
-import VerifyEmail from "./components/VerifyEmail";
-import CompleteOfferHelp from "./components/CompleteOfferHelp";
-import NotifyMe from './components/NotifyMe';
+import VerifyEmail from "./views/VerifyEmail";
+import CompleteOfferHelp from "./views/CompleteOfferHelp";
+import NotifyMe from './views/NotifyMe';
 import ScrollToTop from "./components/ScrollToTop";
 import ShareButtons from "./components/ShareButtons";
 import Presse from './views/Presse';
