+++ src/views/Main.js
@@ -1,8 +1,8 @@
 import { Link } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import '../styles/App.css';
-import FilteredList from './FilteredList';
-import Footer from './Footer';
+import FilteredList from '../components/FilteredList';
+import Footer from '../components/Footer';
 import fb from '../firebase';
 
 export default function Main () {
